requirements.txt
amqp==5.0.1
appdirs==1.4.4
asgiref==3.2.10
billiard==3.6.3.0
boto3==1.15.18
botocore==1.18.18
celery==5.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-storages==1.10.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
importlib-resources==3.0.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==8.0.0
pipenv==2018.11.26
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
virtualenv==20.0.35
virtualenv-clone==0.5.4
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.3.1
